refactor: 픽드랍 섹션을 컴포넌트로 분리
